+++ src/main/java/net/silentchaos512/gear/api/event/GetTraitsEvent.java
@@ -1,7 +1,6 @@
 package net.silentchaos512.gear.api.event;
 
 import net.minecraft.item.ItemStack;
-import net.minecraftforge.eventbus.api.Event;
 import net.silentchaos512.gear.api.parts.PartDataList;
 import net.silentchaos512.gear.api.traits.ITrait;
 
@@ -14,28 +13,12 @@ import java.util.Map;
  * @since 1.3.1
  */
 @SuppressWarnings("AssignmentOrReturnOfFieldWithMutableType")
-public class GetTraitsEvent extends Event {
-    private final ItemStack gear;
-    private final PartDataList parts;
+public class GetTraitsEvent extends GearItemEvent {
     private final Map<ITrait, Integer> traits;
 
     public GetTraitsEvent(ItemStack gear, PartDataList parts, Map<ITrait, Integer> traits) {
-        this.gear = gear;
-        this.parts = parts;
+        super(gear, parts);
         this.traits = traits;
-    }
-
-    @Override
-    public boolean isCancelable() {
-        return false;
-    }
-
-    public ItemStack getGear() {
-        return gear;
-    }
-
-    public PartDataList getParts() {
-        return parts;
     }
 
     public Map<ITrait, Integer> getTraits() {